Refactor driver update workflow helpers
