+++ src/heart/manage/driver_update/exceptions.py
@@ -1,0 +1,5 @@
+"""Driver update error definitions."""
+
+
+class UpdateError(Exception):
+    """Raise when driver update preconditions or downloads fail."""
